+++ src/koptim/__init__.py
@@ -18,7 +18,15 @@ Quickstart::
 
 from koptim._version import __version__
 from koptim.adafusion import Adafusion
+from koptim.adaptive_adafusion import AdafusionProdigy, AdaptiveAdafusion
 from koptim.kprodigy import KProdigy
 from koptim.muon import Muon
 
-__all__ = ["Adafusion", "KProdigy", "Muon", "__version__"]
+__all__ = [
+    "Adafusion",
+    "AdafusionProdigy",
+    "AdaptiveAdafusion",
+    "KProdigy",
+    "Muon",
+    "__version__",
+]